Optimizando y corrigiendo
- Solucionado el fallo de la victoria
- Mejorada la distribución del codigo con el uso de mixins
- Bloqueada la colocación de  fichas  cuando la casilla está ocupada
