+++ othello-vuecli/src/App.vue
@@ -12,9 +12,13 @@
 </template>
 
 <script>
+// Componentes
 import Tablero from "./components/Tablero";
 import Contador from "./components/Contador";
-//import _ from 'lodash';
+// Mixins
+import explorarTablero from "./mixins/explorarTablero";
+import contarFichas from "./mixins/contarFichas";
+
 
 export default {
   name: 'App',
@@ -22,38 +26,17 @@ export default {
     Tablero,
     Contador
   },
-  methods: {
-  },
   computed: {
+    // Jugador Activo
     jugadorActivoLocal: function () {
       // Traigo localmente el estado del jugador activo
       return this.$store.state.jugadorActivo
-    },
-    fichasNegras: function () {
-      // Creo un array para almacenar las fichas negras
-      let fichas = [];
-      // Filtro el array para quedarme las fichas negras
-      fichas = this.$store.getters.tablero.filter(casilla => {
-        return casilla === this.$store.state.fichaNegra;
-      });
-      // Devuelvo el total de posiciones
-      return fichas.length;
-    },
-    fichasBlancas: function () {
-      // Creo un array para almacenar las fichas negras
-      let fichas = [];
-      // Filtro el array para quedarme las fichas negras
-      fichas = this.$store.getters.tablero.filter(casilla => {
-        return casilla === this.$store.state.fichaBlanca;
-      });
-      // Devuelvo el total de posiciones
-      return fichas.length;
-    },
+    }
   },
   watch: {
     jugadorActivoLocal: function () {
       // Observo el jugador activo para analizar el estado del tablero, si no hay casillas vacias entro
-      if (this.$store.state.victoria === true) {
+      if (this.$store.getters.tablero.includes(!this.$store.state.casillaVacia)) {
         // Voy a la vista de victoria
         this.$router.push('Victoria')
       }
@@ -61,8 +44,14 @@ export default {
       console.log(this.jugadorActivoLocal);
       console.log('Negras '.concat(this.fichasNegras));
       console.log('Blancas '.concat(this.fichasBlancas));
-    }
-  }
+    },
+  },
+  mixins: [
+    // Methods
+    explorarTablero,
+    // Computed
+    contarFichas
+  ],
 }
 </script>
