+++ othello-vuecli/src/views/Victoria.vue
@@ -2,9 +2,9 @@
   <div class="menus">
     <div class="menu">
       <!--- Cabecera del menu --->
-      <CabeceraMenus v-if="fichasNegrasLocal < fichasBlancasLocal" class="menu__titulo--especial" texto="ganan blancas"/>
-      <CabeceraMenus v-if="fichasNegrasLocal > fichasBlancasLocal" texto="ganan negras"/>
-      <CabeceraMenus v-if="fichasNegrasLocal === fichasBlancasLocal" texto="Empate"/>
+      <CabeceraMenus v-if="fichasNegras < fichasBlancas" class="menu__titulo--especial" texto="ganan blancas"/>
+      <CabeceraMenus v-if="fichasNegras > fichasBlancas" texto="ganan negras"/>
+      <CabeceraMenus v-if="fichasNegras === fichasBlancas" texto="Empate"/>
       <nav class="menu__nav">
         <ul class="menu__nav__ul">
           <!--- Botones del menu --->
@@ -18,10 +18,12 @@
 </template>
 
 <script>
-// @ is an alias to /src
+// Componentes
 import BotonJuego from "../components/botones/BotonJuego";
 import BotonVolverAInicio from "../components/botones/BotonVolverAInicio";
 import CabeceraMenus from "../components/textos/CabeceraMenus";
+// Mixins
+import contarFichas from "../mixins/contarFichas";
 
 export default {
   name: 'Victoria',
@@ -30,27 +32,9 @@ export default {
     BotonVolverAInicio,
     BotonJuego,
   },
-  computed: {
-    fichasNegrasLocal: function () {
-      // Creo un array para almacenar las fichas negras
-      let fichas = [];
-      // Filtro el array para quedarme las fichas negras
-      fichas = this.$store.getters.tablero.filter(casilla => {
-        return casilla === this.$store.state.fichaNegra;
-      });
-      // Devuelvo el total de posiciones
-      return fichas.length;
-    },
-    fichasBlancasLocal: function () {
-      // Creo un array para almacenar las fichas negras
-      let fichas = [];
-      // Filtro el array para quedarme las fichas negras
-      fichas = this.$store.getters.tablero.filter(casilla => {
-        return casilla === this.$store.state.fichaBlanca;
-      });
-      // Devuelvo el total de posiciones
-      return fichas.length;
-    }
-  }
+  mixins: [
+    // Computed
+    contarFichas
+  ]
 }
 </script>
